bubble_sort/binary_search/indexes
Inputs row of numbers, makes a list of them, sorts low to high, asks for random number and searches if the list has this random number. If the list includes the random number, the program reports and shows both previous and given numbers indexes in the list. If there is no such number in the list, the program also reports and shows the closest numbers indexes in the list.